requirements.txt
asgiref==3.3.4
Django==3.2.3
django-crispy-forms==1.11.2
django-filter==2.4.0
django-tables2==2.3.4
pytz==2021.1
sqlparse==0.4.1
